requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
openai==1.3.0
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
textblob==0.17.1
drf-yasg==1.21.10